move serialization-core to api
