+++ api/src/main/java/net/automatalib/serialization/InputModelData.java
