+++ api/src/main/java/net/automatalib/serialization/InputModelSerializer.java
