+++ api/src/main/java/net/automatalib/serialization/ModelSerializer.java
