+++ api/src/main/java/net/automatalib/serialization/automaton/SimpleAutomatonSerializationProvider.java
