+++ api/src/main/java/net/automatalib/serialization/automaton/SimpleAutomatonSerializer.java
