+++ pom.xml
@@ -387,11 +387,6 @@ limitations under the License.
             </dependency>
             <dependency>
                 <groupId>net.automatalib</groupId>
-                <artifactId>automata-serialization-core</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>net.automatalib</groupId>
                 <artifactId>automata-serialization-dot</artifactId>
                 <version>${project.version}</version>
             </dependency>
