+++ serialization/aut/src/main/java/module-info.java
@@ -32,7 +32,6 @@ open module net.automatalib.serialization.aut {
     requires net.automatalib.api;
     requires net.automatalib.common.util;
     requires net.automatalib.core;
-    requires net.automatalib.serialization.core;
 
     exports net.automatalib.serialization.aut;
 }
