--- serialization/core/pom.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-Copyright (C) 2013-2024 TU Dortmund University
-This file is part of AutomataLib, http://www.automatalib.net/.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>net.automatalib</groupId>
-        <artifactId>automata-serialization-parent</artifactId>
-        <version>0.12.0-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <artifactId>automata-serialization-core</artifactId>
-    <packaging>jar</packaging>
-
-    <name>AutomataLib :: Serialization :: Core</name>
-    <description>Generic automata serialization core</description>
-
-    <dependencies>
-        <!-- internal -->
-        <dependency>
-            <groupId>net.automatalib</groupId>
-            <artifactId>automata-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.automatalib</groupId>
-            <artifactId>automata-commons-util</artifactId>
-        </dependency>
-    </dependencies>
-</project>
