+++ serialization/dot/src/main/java/module-info.java
@@ -32,7 +32,6 @@ open module net.automatalib.serialization.dot {
     requires net.automatalib.api;
     requires net.automatalib.common.util;
     requires net.automatalib.core;
-    requires net.automatalib.serialization.core;
     requires org.checkerframework.checker.qual;
 
     exports net.automatalib.serialization.dot;
