+++ serialization/dot/src/test/java/net/automatalib/serialization/dot/DOTDeserializationTest.java
@@ -39,8 +39,8 @@ import net.automatalib.automaton.transducer.MooreMachine;
 import net.automatalib.automaton.transducer.impl.CompactMealy;
 import net.automatalib.automaton.transducer.impl.CompactMoore;
 import net.automatalib.common.util.io.UnclosableInputStream;
+import net.automatalib.exception.FormatException;
 import net.automatalib.graph.UniversalGraph;
-import net.automatalib.serialization.FormatException;
 import net.automatalib.serialization.InputModelData;
 import net.automatalib.ts.modal.impl.CompactMTS;
 import net.automatalib.word.Word;
