+++ serialization/etf/pom.xml
@@ -41,10 +41,6 @@ limitations under the License.
             <groupId>net.automatalib</groupId>
             <artifactId>automata-commons-util</artifactId>
         </dependency>
-        <dependency>
-            <groupId>net.automatalib</groupId>
-            <artifactId>automata-serialization-core</artifactId>
-        </dependency>
 
         <!-- external -->
         <dependency>
