+++ serialization/etf/src/main/java/module-info.java
@@ -31,7 +31,6 @@ open module net.automatalib.serialization.etf {
     requires com.google.common;
     requires net.automatalib.api;
     requires net.automatalib.common.util;
-    requires net.automatalib.serialization.core;
 
     exports net.automatalib.serialization.etf.writer;
 }
