+++ serialization/learnlibv2/pom.xml
@@ -47,10 +47,6 @@ limitations under the License.
         </dependency>
         <dependency>
             <groupId>net.automatalib</groupId>
-            <artifactId>automata-serialization-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.automatalib</groupId>
             <artifactId>automata-util</artifactId>
         </dependency>
