+++ serialization/learnlibv2/src/main/java/module-info.java
@@ -32,7 +32,6 @@ open module net.automatalib.serialization.learnlibv2 {
     requires net.automatalib.api;
     requires net.automatalib.common.util;
     requires net.automatalib.core;
-    requires net.automatalib.serialization.core;
     requires net.automatalib.util;
 
     exports net.automatalib.serialization.learnlibv2;
