+++ serialization/pom.xml
@@ -33,7 +33,6 @@ limitations under the License.
 
     <modules>
         <module>aut</module>
-        <module>core</module>
         <module>dot</module>
         <module>etf</module>
         <module>fsm</module>
