+++ serialization/saf/src/main/java/module-info.java
@@ -31,7 +31,6 @@ open module net.automatalib.serialization.saf {
     requires net.automatalib.api;
     requires net.automatalib.common.util;
     requires net.automatalib.core;
-    requires net.automatalib.serialization.core;
     requires org.checkerframework.checker.qual;
 
     exports net.automatalib.serialization.saf;
