+++ serialization/taf/pom.xml
@@ -45,10 +45,6 @@ limitations under the License.
             <groupId>net.automatalib</groupId>
             <artifactId>automata-core</artifactId>
         </dependency>
-        <dependency>
-            <groupId>net.automatalib</groupId>
-            <artifactId>automata-serialization-core</artifactId>
-        </dependency>
 
         <!-- build -->
         <dependency>
